+++ denoiser/modules.py
@@ -1,5 +1,10 @@
 import torch
 from torch import nn
+from denoiser.improve_single_trans import TransformerEncoderLayer
+from torch.nn.utils import weight_norm
+from .resample import downsample2, upsample2
+
+from .utils import capture_init
 
 class BLSTM(nn.Module):
     def __init__(self, dim, layers=2, bi=True):
@@ -135,3 +140,109 @@ class PixelUnshuffle1D(torch.nn.Module):
         x = x.permute(0, 3, 1, 2).contiguous()
         x = x.view([batch_size, short_channel_len, short_width])
         return x
+
+
+def WNConv1d(*args, **kwargs):
+    return weight_norm(nn.Conv1d(*args, **kwargs))
+
+
+def weights_init(m):
+    classname = m.__class__.__name__
+    if classname.find("Conv") != -1:
+        m.weight.data.normal_(0.0, 0.02)
+    elif classname.find("BatchNorm2d") != -1:
+        m.weight.data.normal_(1.0, 0.02)
+        m.bias.data.fill_(0)
+
+
+def count_parameters(model):
+    return sum(p.numel() for p in model.parameters() if p.requires_grad)
+
+
+class NLayerDiscriminator(nn.Module):
+    def __init__(self, ndf, n_layers, downsampling_factor):
+        super().__init__()
+        model = nn.ModuleDict()
+
+        model["layer_0"] = nn.Sequential(
+            nn.ReflectionPad1d(7),
+            WNConv1d(1, ndf, kernel_size=15),
+            nn.LeakyReLU(0.2, True),
+        )
+
+        nf = ndf
+        stride = downsampling_factor
+        for n in range(1, n_layers + 1):
+            nf_prev = nf
+            nf = min(nf * stride, 1024)
+
+            model["layer_%d" % n] = nn.Sequential(
+                WNConv1d(
+                    nf_prev,
+                    nf,
+                    kernel_size=stride * 10 + 1,
+                    stride=stride,
+                    padding=stride * 5,
+                    groups=nf_prev // 4,
+                ),
+                nn.LeakyReLU(0.2, True),
+            )
+
+        nf = min(nf * 2, 1024)
+        model["layer_%d" % (n_layers + 1)] = nn.Sequential(
+            WNConv1d(nf_prev, nf, kernel_size=5, stride=1, padding=2),
+            nn.LeakyReLU(0.2, True),
+        )
+
+        model["layer_%d" % (n_layers + 2)] = WNConv1d(
+            nf, 1, kernel_size=3, stride=1, padding=1
+        )
+
+        self.model = model
+
+    def forward(self, x):
+        results = []
+        for key, layer in self.model.items():
+            x = layer(x)
+            results.append(x)
+        return results
+
+
+class Discriminator(nn.Module):
+    @capture_init
+    def __init__(self, num_D, ndf, n_layers, downsampling_factor):
+        super().__init__()
+        self.model = nn.ModuleDict()
+        self.num_D = num_D
+        for i in range(num_D):
+            self.model[f"disc_{i}"] = NLayerDiscriminator(
+                ndf, n_layers, downsampling_factor
+            )
+
+        self.downsample = nn.AvgPool1d(4, stride=2, padding=1, count_include_pad=False)
+        self.apply(weights_init)
+
+    def forward(self, x):
+        results = []
+        for key, disc in self.model.items():
+            results.append(disc(x))
+            x = self.downsample(x)
+        return results
+
+
+class LaplacianDiscriminator(Discriminator):
+
+    def forward(self, x):
+        results = []
+        for i, (key, disc) in enumerate(self.model.items()):
+            if i == 0:  # insert total frequency range
+                results.append(disc(x))
+            elif i == self.num_D - 1:  # insert 0 - total frequency range / 2^num_D
+                downsampled_x = downsample2(x)
+                results.append(disc(downsampled_x))
+            else:  # insert total_frequency_range / 2^i - total_frequency_range/2^(i-1)
+                downsampled_x = downsample2(x)
+                laplacian = x - upsample2(downsampled_x)
+                results.append(disc(laplacian))
+                x = downsampled_x
+        return results
